The 25015 is gone, and ota/publish.sh replaces the chip-tool update path
No 25015 board exists in the house, so its board definition, partition map,
overlays and the conditions that mentioned it are removed, along with the
docs that told a reader to build for it.

ota/update.sh drove updates with chip-tool on a fabric that no longer exists.
publish.sh does what was being done by hand: a pristine build, the version
read back from the signed image, the descriptor matter-server wants written
next to it on the Pi, and a restart so it is read. The panel starts each
update.

Also: two narrowing warnings in status_led.cpp, and a [hidden] rule for the
panel - the colour wheel was hidden on white bulbs and still drawn, because
the attribute lost to the class's display: flex.

+++ firmware/src/status_led.cpp
@@ -340,7 +340,7 @@ void SetBackground(Pattern p, Rgb color, uint16_t periodMs)
 
 void SetBackgroundScaled(Pattern p, Rgb color, uint8_t maxScale, uint16_t periodMs)
 {
-	Effect next = { p, color, periodMs ? periodMs : 1000, maxScale };
+	Effect next = { p, color, static_cast<uint16_t>(periodMs ? periodMs : 1000), maxScale };
 
 	/* If nothing changed, do not restart the phase. Otherwise the indicator,
 	 * refreshed once a minute, would restart the pulse and look irregular. */
@@ -356,7 +356,7 @@ void SetBackgroundScaled(Pattern p, Rgb color, uint8_t maxScale, uint16_t period
 
 void SetOverlay(Pattern p, Rgb color, uint16_t periodMs)
 {
-	sOverlay = { p, color, periodMs ? periodMs : 1000, 255 };
+	sOverlay = { p, color, static_cast<uint16_t>(periodMs ? periodMs : 1000), 255 };
 	sOverlayActive = true;
 	sPhaseStart = k_uptime_get();
 	Reschedule();
